backend/requirements.txt
pydantic~=2.10.4
uvicorn~=0.33.0
fastapi~=0.115.6
sse-starlette~=2.2.1
starlette~=0.44.0
bson~=0.5.10
pymongo~=4.10.1
python-dotenv~=1.0.1
motor~=3.6.0
tqdm~=4.67.1